fix bug №7 rewrite p_norm with power function 4 dec 2012

+++ src/p-adic.h
@@ -69,11 +69,11 @@ pa_num* p_gamma_pa_num(pa_num *pa, int gamma);
 
 void print_pa_num(pa_num *pa);
 
-float p_norm(pa_num *pa);
+double p_norm(pa_num *pa);
 
 int indicator(pa_num *x, pa_num *n, int gamma);
 
-float from_canonic_to_float(pa_num *pa);
+double from_canonic_to_float(pa_num *pa);
 
 pa_num* get_fractional_part(pa_num *pa);
 
@@ -83,9 +83,9 @@ complex wavelet(pa_num *x, pa_num *n, int gamma, int j);
 
 pa_num* jmult(pa_num *pa1, int j);
 
-float integral(float (*func)(pa_num* pnum), int g_min, int g_max);
+double integral(double (*func)(pa_num* pnum), int g_min, int g_max);
 
-complex wavelet_integral(float (*func)(pa_num *pnum), pa_num *n, int gamma, int j, int g_min, int g_max);
+complex wavelet_integral(double (*func)(pa_num *pnum), pa_num *n, int gamma, int j, int g_min, int g_max);
 
 /* Not implemented yet */
 pa_num* mult(pa_num *pa1, pa_num *pa2);
